moved doc scss and js into components folder.
