+++ app/templates/src/docs/components/card/_card--component.scss
