+++ app/templates/src/docs/components/card/_card--rich-text.scss
