+++ app/templates/src/docs/components/card/_card.scss
