+++ app/templates/src/docs/components/color-swatch/_color-swatch.scss
