+++ app/templates/src/docs/components/component-view/_component-view.js
