+++ app/templates/src/docs/components/nav/_nav.scss
