+++ app/templates/src/docs/components/perspective/_perspective.scss
