+++ app/templates/src/docs/components/resizer/_resizer.scss
