+++ app/templates/src/docs/components/tabs/_tabs.scss
