+++ app/templates/src/docs/components/template/_template.scss
