+++ app/templates/src/docs/scss/docs.scss
@@ -7,20 +7,20 @@ $breakpoint--lg: min-width 1200px; // desktop;
 
 $border-radius: 2px;
 
-@import './prism'; // syntax color coding;
+@import '../components/prism/prism'; // syntax color coding;
 @import './base'; // global styles;
-@import './button';
-@import './card';
-@import './card--colors';
-@import './card--component';
-@import './card--font';
-@import './card--rich-text';
-@import './color-swatch';
-@import './component-view';
-@import './grid';
-@import './nav';
-@import './page-header';
-@import './perspective';
-@import './resizer';
-@import './tabs';
-@import './template';
+@import '../components/button/button';
+@import '../components/card/card';
+@import '../components/card/card--colors';
+@import '../components/card/card--component';
+@import '../components/card/card--font';
+@import '../components/card/card--rich-text';
+@import '../components/color-swatch/color-swatch';
+@import '../components/component-view/component-view';
+@import '../components/grid/grid';
+@import '../components/nav/nav';
+@import '../components/page-header/page-header';
+@import '../components/perspective/perspective';
+@import '../components/resizer/resizer';
+@import '../components/tabs/tabs';
+@import '../components/template/template';
